refactor: Remove unnecessary 'required_live_plan' key in exchange_info
This commit removes the unnecessary 'required_live_plan' key in the 'exchange_info' dictionary in 'jesse/info.py'. The 'required_live_plan' key was set to 'premium' for all exchanges, but it is no longer needed as it does not affect the functionality of the code. Removing it improves code cleanliness and reduces unnecessary complexity.